ok we have cr-sqlite installed on the table and the ap doesnt crash!
now to fix the connect issues when that requery seems to have when we
are not on the react native view

+++ index.tsx
@@ -15,8 +15,9 @@ console.log(MyModuleView);
 console.log(setValueAsync);
 
 const HelloWorld = () => {
-  const {data : psEvents } = useQuery<string>('select * from eventsV9'); 
+  const {data : psEvents } = useQuery<string>('select id, nid, ttl from eventsV9'); 
   const [sqliteEvents, setSqliteEvents] = useState<any[]>([]);
+  const [tempTableEvents, setTempTableEvents] = useState<any[]>([]);
   const [dbStatus, setDbStatus] = useState<string>('');
   const [lastUpdate, setLastUpdate] = useState<string>('');
 
@@ -28,28 +29,48 @@ const HelloWorld = () => {
       try {
         await regDb.execute("SELECT load_extension('crsqlite')");
         
-        // Unique index on 2 columns gets error below
-        // statement execution error: Table eventsV9 has unique indices besidesthe primary key. This is not allowed for CRRs
-        await regDb.execute("DROP INDEX IF EXISTS eventsIdxV9");
+        // Check if eventsV9 exists before creating temp table
+        const tableExists = await regDb.execute("SELECT name FROM sqlite_master WHERE type='table' AND name='eventsV9'");
+        console.log('tableExists', tableExists?.rows);
 
-        // add primary key
-        await regDb.execute("CREATE TABLE IF NOT EXISTS eventsV9_temp AS SELECT * FROM eventsV9");
-        let result = await regDb.execute("select id from eventsV9_temp");
-        console.log(result?.rows);
+        if (tableExists?.rows?.length > 0) {
+            // Unique index on 2 columns gets error below
+            // statement execution error: Table eventsV9 has unique indices besides the primary key. This is not allowed for CRRs
+            await regDb.execute("DROP INDEX IF EXISTS eventsIdxV9");
+            await regDb.execute("CREATE TABLE IF NOT EXISTS eventsV9_temp AS SELECT * FROM eventsV9");
 
-        result = await regDb.execute("select id, Count(*) from eventsV9_temp group by id");
-        console.log(result?.rows);
+            const result = await regDb.execute("select id, Count(*) from eventsV9_temp group by id");
+            console.log("num events in temp table:", result?.rows?.length);
 
-        // await regDb.execute("DROP TABLE eventsV9");
-        await regDb.execute("CREATE TABLE IF NOT EXISTS eventsV9  (id INTEGER NOT NULL PRIMARY KEY, cid INTEGER, rep INTEGER, alld INTEGER, nid INTEGER, ttl TEXT, s1 TEXT, estart INTEGER, eend INTEGER, istart INTEGER, iend INTEGER, loc TEXT, snz INTEGER, dsts INTEGER, ls INTEGER, clr INTEGER, altm INTEGER, ogn INTEGER, fsn INTEGER, attsts INTEGER, oattsts INTEGER, i1 INTEGER, i2 INTEGER, i3 INTEGER, i4 INTEGER, i5 INTEGER, i6 INTEGER, i7 INTEGER, i8 INTEGER, s2 TEXT)");
-        await regDb.execute("INSERT OR IGNORE INTO eventsV9 SELECT * FROM eventsV9_temp");
-        // await regDb.execute("DROP TABLE eventsV9_temp");
+            // Get schema for eventsV9 
+            const eventsV9Schema = await regDb.execute("SELECT sql FROM sqlite_master WHERE type='table' AND name='eventsV9'");
+            console.log('original eventsV9 schema:', eventsV9Schema?.rows);
 
+            const newSchemaStatementLines = (eventsV9Schema?.rows?.[0]?.sql as string).split('\n')
+
+            const idIndex = newSchemaStatementLines.findIndex(i => i === "  id INT," || i === "  id INTEGER,")
+            newSchemaStatementLines[idIndex] = "  id INT NOT NULL PRIMARY KEY,"
+
+            const newSchema = newSchemaStatementLines.join('\n')
+            .replace(", PRIMARY KEY (id, istart)", "")
+
+            console.log('new Crsql compatible schema:', newSchema);
+            
+            // Drop eventsV9 table with old schema
+            await regDb.execute("DROP TABLE IF EXISTS eventsV9");
+
+            // Create eventsV9 table with schema 
+            await regDb.execute(newSchema);
+            
+            try {
+              await regDb.execute(`INSERT OR REPLACE INTO eventsV9 SELECT * FROM eventsV9_temp`);
+            } catch (error) {
+              console.error('Failed to insert/update data from temp table:', error);
+            }
+        } 
+        
         // recreate index with no unique constraint
         await regDb.execute("CREATE INDEX IF NOT EXISTS eventsIdxV9 ON eventsV9 (id, istart)");
-        
-        // Initialize cr-sqlite
-        // await regDb.execute("SELECT crsql_finalize()");
         
         // Enable CRDT behavior for the table
         await regDb.execute("SELECT crsql_as_crr('eventsV9')");
@@ -59,10 +80,16 @@ const HelloWorld = () => {
 
       
       // Query events from SQLite
-      const result = await regDb.execute('SELECT * FROM eventsV9 limit 1');
+      const result = await regDb.execute('SELECT id, nid, ttl FROM eventsV9');
       if (result?.rows) {
         setSqliteEvents(result.rows || []);
       }
+        // Query temp table events
+        const tempResult = await regDb.execute('SELECT id, nid, ttl FROM eventsV9_temp');
+        if (tempResult?.rows) {
+          setTempTableEvents(tempResult.rows || []);
+        }
+
     })();
 
     addChangeListener((value) => {
@@ -87,7 +114,9 @@ const HelloWorld = () => {
       <Text style={styles.hello}>PowerSync Status: {dbStatus}</Text>
       <Text style={styles.hello}>Last Updated: {lastUpdate}</Text>
       <Text style={styles.hello}>PowerSync Events: {JSON.stringify(psEvents)}</Text>
-      <Text style={styles.hello}>SQLite Events: {JSON.stringify(sqliteEvents)}</Text>
+
+      <Text style={styles.hello}>SQLite Events eventsV9: {JSON.stringify(sqliteEvents)}</Text>
+      <Text style={styles.hello}>Events V9 Temp Table: {JSON.stringify(tempTableEvents)}</Text>
       <MyModuleView name="MyModuleView" />
       <Button
         title="Click me!"
